add auth to main navigation flows (separate auth screen stack from main app stack via parent switch navigator
