+++ navigation/AppNavigator.js
@@ -2,11 +2,16 @@ import React from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 
 import MainTabNavigator from './MainTabNavigator';
+import AuthNavigator from './AuthNavigator';
 
 export default createAppContainer(
   createSwitchNavigator({
     // You could add another route here for authentication.
     // Read more at https://reactnavigation.org/docs/en/auth-flow.html
+    Auth: AuthNavigator,
     Main: MainTabNavigator,
+  },
+  {
+    initialRouteName: 'Auth',
   })
 );
